Reverted to edit edt-08e14f27-5ae5-4673-bbe9-9c4565df52ac: "Implement the plan"
